utils: move Promise and DeepCopy to dependency-free subpackages
Importing utils for a promise costs 497 packages: codec.go reaches pion/webrtc
and 74 packages behind it, and the rest of utils pulls in psrpc and nats. A
consumer whose only use is *utils.Promise on one interface method inherits all
of it, and where that consumer is translated for deterministic simulation, so
is every package in the closure.

utils/promise and utils/deepcopy carry nothing beyond the standard library:
context and sync, and reflect. Their constructors drop the stutter, and
NewResolved is spelled apart from the Resolved method, which reports whether a
promise has settled rather than producing one that has.

The old spellings stay. Promise is a type alias rather than a defined type, so
a promise built under either name satisfies an interface written against the
other — the case that matters when one package declares the seam and another
implements it. A compile-time assertion pins that, since a defined type would
still build here and break only the callers that mix the two.
